# next-client/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WealthFlow API",
        "description": "API documentation for WealthFlow financial tracking application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base URL"
        }
    ],
    "paths": {
        "/expense": {
            "get": {
                "summary": "Get all expenses for a user",
                "description": "Retrieves all expense records for a specific user",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the user to fetch expenses for"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Expense"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new expense",
                "description": "Adds a new expense record to the database",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewExpense"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid expense data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update an existing expense",
                "description": "Updates an existing expense record in the database",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateExpense"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid expense data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete an expense",
                "description": "Deletes an expense record from the database",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the expense to delete"
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the user who owns the expense"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Expense": {
                "type": "object",
                "required": [
                    "id",
                    "userid",
                    "name",
                    "date",
                    "amount",
                    "type",
                    "paymentmethod",
                    "category"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "userid": {
                        "type": "integer",
                        "format": "int64",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Grocery Shopping"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-20T14:00:00Z"
                    },
                    "amount": {
                        "type": "string",
                        "example": "85.75"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "income",
                            "expense"
                        ],
                        "example": "expense"
                    },
                    "paymentmethod": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "credit",
                            "debit"
                        ],
                        "example": "debit"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "Food",
                            "Rent",
                            "Transport",
                            "Entertainment",
                            "Health",
                            "Utilities",
                            "Education",
                            "Shopping",
                            "Other"
                        ],
                        "example": "Food"
                    }
                }
            },
            "NewExpense": {
                "type": "object",
                "required": [
                    "userId",
                    "name",
                    "date",
                    "amount",
                    "type",
                    "paymentMethod",
                    "category"
                ],
                "properties": {
                    "userId": {
                        "type": "integer",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Grocery Shopping"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-20T14:00:00Z"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 85.75
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "income",
                            "expense"
                        ],
                        "example": "expense"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "credit",
                            "debit"
                        ],
                        "example": "debit"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "Food",
                            "Rent",
                            "Transport",
                            "Entertainment",
                            "Health",
                            "Utilities",
                            "Education",
                            "Shopping",
                            "Other"
                        ],
                        "example": "Food"
                    }
                }
            },
            "UpdateExpense": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Grocery Shopping"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-20T14:00:00Z"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 85.75
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "income",
                            "expense"
                        ],
                        "example": "expense"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "credit",
                            "debit"
                        ],
                        "example": "debit"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "Food",
                            "Rent",
                            "Transport",
                            "Entertainment",
                            "Health",
                            "Utilities",
                            "Education",
                            "Shopping",
                            "Other"
                        ],
                        "example": "Food"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "error": {
                        "type": "object",
                        "example": {}
                    }
                }
            }
        }
    }
}